EEPROM & STP
Changed handling of STP/NPT changes for better backwards compatibility.
EEPROM saving for gas selection and setpoints is now toggleable. If either is changed a number of times beyond a default value, EEPROM saving will be disabled in order to preserve the chip as it is good for about 1 million write operations per register.